(PC-17376)[ADAGE] feat: move functions into utils directory
