+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/Offers/utils/extractOfferIdFromObjectId.ts
@@ -1,0 +1,9 @@
+export const extractOfferIdFromObjectId = (offerId: string): number => {
+  const splitResult = offerId.split('T-')
+
+  if (splitResult.length === 2) {
+    return Number(splitResult[1])
+  }
+
+  return Number(offerId)
+}
