+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/Offers/utils/index.ts
@@ -1,0 +1,5 @@
+export * from './extractOfferIdFromObjectId'
+export * from './formatDescription'
+export * from './getOfferVenueAndOffererName'
+export * from './offerIsBookable'
+export * from './offerIsCollectiveOffer'
