+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/Offers/utils/offerIsBookable.ts
@@ -1,0 +1,8 @@
+import {
+  CollectiveOfferResponseModel,
+  CollectiveOfferTemplateResponseModel,
+} from 'apiClient'
+
+export const offerIsBookable = (
+  offer: CollectiveOfferResponseModel | CollectiveOfferTemplateResponseModel
+): boolean => !offer.isSoldOut && !offer.isExpired
